refactor: Move template/ to internal
